Move config stuff out of main
This moves config stuff out of main.go and adds db stuff to the config file. Because encapsulation.
