+++ main.go
@@ -1,53 +1,19 @@
 package main
 
 import (
-	"os"
-
 	"github.com/eputnam/health-check-server/api"
+	"github.com/eputnam/health-check-server/config"
 	"github.com/eputnam/health-check-server/db"
-	"gopkg.in/yaml.v2"
 )
 
-const (
-	config_path = "config.yaml"
-)
-
-var appConfig AppConfig
-
-func checkError(err error) {
-	if err != nil {
-		panic(err)
-	}
-}
-
-/*
-CONFIG STUFF
-*/
-type AppConfig struct {
-	Server struct {
-		Port string `yaml:"port"`
-	} `yaml:"server"`
-}
-
-func loadConfig() AppConfig {
-	file, err := os.Open(config_path)
-	checkError(err)
-	defer file.Close()
-
-	var config AppConfig
-	decoder := yaml.NewDecoder(file)
-	err = decoder.Decode(&config)
-	checkError(err)
-
-	return config
-}
+var globalConfig config.GlobalConfig
 
 func init() {
-	appConfig = loadConfig()
+	globalConfig = config.LoadConfig()
 }
 
 func main() {
-	store := db.NewStore()
+	store := db.NewStore(globalConfig)
 	server := api.NewServer(store)
-	server.StartServer("localhost:" + appConfig.Server.Port)
+	server.StartServer(globalConfig.Server.Host + ":" + globalConfig.Server.Port)
 }
